+++ tests/integration/Driver/RedisQueueCest.php
@@ -3,9 +3,9 @@
 namespace Tests\Integration\Driver;
 
 use Initx\Querabilis\Driver\RedisQueue;
-use Tests\Double\EnvelopeMother;
-use Tests\Double\PredisClientMother;
-use Tests\IntegrationTester;
+use Initx\Querabilis\Tests\Double\EnvelopeMother;
+use Initx\Querabilis\Tests\Double\PredisClientMother;
+use Initx\Querabilis\Tests\IntegrationTester;
 
 class RedisQueueCest
 {